Update README.md
Redact a small syntax error, and expand the comment for clarity as to when to expect the update information to show.
n.b. perhaps this should also be a note, instead of a warning? Your judgment, and not in any way crucial...on this great vuepress system, for which thanks.