fix(dev): Stop observing add/remove of files in node_modules
Closes #855

When running vuepress dev, chorkidar is set to observe all
addition/removal of md files inside the entity folder.

`node_modules` is one of these folders and can contain tons of
subfolders. It mostly never contains any md files anyway, and
even rarely has any hot additions of md files.

Another perk of this perf fix is this: If for some reason a developer
decided to publish a folder with a weird name in its module. It will not
prevent `vuepress dev` from loading.