Added trailing forward-slashes in link properties
The docs show how to have sub groups inside a dropdown menu by having nested items. However, the example link text for Chinese and Japanese does not work as expected.  Add the necessary trailing forward-slashes so that the code works as expected if somebody was to cut-and-paste the example as is.