Add CSS preprocessor tip to Build Pipeline section
While this document is solely concerned with documenting the configuration options, it isn't very explicit regarding structure or the addition of dependencies such as the CSS pre-processors. This change attempts to clarify that some by pointing to more instructive sections of the site better detailing how to configure CSS pre-processors.