Added shared assumption config.js changed netlify
When I was going through the deployment guide. I didn't understand what the config.js was supposed to look like. It took a bit of searching to be able to deploy successfully.

Also I changed the netlify build command, because sometimes the `yarn docs:build` or `npm run docs:build` wouldn't work or would fail, but `vuepress build docs` seems to work consistently.