feat($plugin-search): prevent iOS input zoom
Apple decided that form inputs on iOS need to be a comfortable reading size, set at a minimum of 16px. If the form input font-size is less than that, they will automatically zoom into the field until the text appears at that size.

To prevent this behaviour this commit sets the font-size to 16px.